Made buildTopDownMergedConfiguration really recursive - in most cases it didn't.
Fixed a merge problem where only direct children of <configuration> in the POM
appeared in the merged configuration.

(like maven-antrun-plugin, the 'tasks' configuration tag didn't have any
value, but has children. After the merge, the children were gone,
and it was marked as having no value, and hence the field was not updated.)


git-svn-id: https://svn.apache.org/repos/asf/maven/components/trunk/maven-core@232281 13f79535-47bb-0310-9956-ffa450edef68