Resolving: MNG-449, MNG-832
o If the metadata version is still in the plugin artifact after it's resolved, then simply return null and allow the plugin version resolution process to fail.

o Added a new mojo to update the plugin registry (if it's enabled) with newly installed versions of plugins.

o Added comments to maven-settings and maven-plugin-parameter-documenter poms marking the packageWithVersion config as deprecated pending modello alpha-5

o Cleaned up it0013 and it0020 to work with the new resolution (point one above).



git-svn-id: https://svn.apache.org/repos/asf/maven/components/trunk/maven-core@293520 13f79535-47bb-0310-9956-ffa450edef68