PR: MNG-1525 Submitted By: Edwin Punzalan Reviewed By: John Casey
NOT applying this patch. I found a better solution that will factor the interpolation of the POM into a flexible utility in plexus-utils, and will allow introduction of envar resolution to the POM. It will also make interpolating the settings.xml and profiles.xml using any of a number of expression resolvers (using envar resolution only for now).

BTW, I tried using System.getenv(..) in JDK1.4, and it fails with java.lang.Error and a deprecation message. So, I'm using Edwin's code to extract the envars into a Properties object. We can improve this later.


git-svn-id: https://svn.apache.org/repos/asf/maven/components/trunk/maven-project@354462 13f79535-47bb-0310-9956-ffa450edef68