PR: MNG-1630 Submitted By: Edwin Punzalan Reviewed By: John Casey
Applied patches, with minor changes.

These patches will ensure that the optional flag is passed on and inherited correctly when dealing with managed dependencies.

I changed the patches, in that I added a new createDependencyArtifact(..) method on ArtifactFactory, which will eliminate the need to call an older variant of the method by passing in a null value for the inheritedScope parameter.


git-svn-id: https://svn.apache.org/repos/asf/maven/components/trunk/maven-project@354544 13f79535-47bb-0310-9956-ffa450edef68