PR: MNG-1550 Submitted By: Edwin Punzalan Reviewed By: John Casey
Applied patch, with small change. This patch will save the checksums in the local repo, once they've transferred safely. The change I made was to pass in the "real" destination to the verifyChecksum method, rather than only the temp destination. This will allow the checksums to actually persist in the local repo, rather than a temporary directory.



git-svn-id: https://svn.apache.org/repos/asf/maven/components/trunk@354631 13f79535-47bb-0310-9956-ffa450edef68