PR: MNG-1471 Submitted By: Edwin Punzalan Reviewed By: John Casey
Applied patch, with small logic fix. This patch will validate that if a systemPath is specified for a dependency, it is absolute. It's possible to use expressions in this path, but by the time it's validated in this step, those expressions will have been resolved.

My change was to add the same validation to managed dependencies.



git-svn-id: https://svn.apache.org/repos/asf/maven/components/trunk/maven-project@354635 13f79535-47bb-0310-9956-ffa450edef68