o change the assembly so that classworlds is separated from the rest of the uber jar, otherwise everything is jammed in the   system classloader which is crap.
MNG-2894 Hiding the use of plexus-utils in the core by rewriting all references

MNG-2897 We have to let Xpp3Dom be in the core in its original form because it is used in plugins. This was a big mistake. In 2.1
         the DOM used for plugin configuration must be in the plugin API.

MNG-2894 Check to see if a plugin uses p-u, if it's less then 1.1 bump it up, and if it doesn't specify the give it 1.1. There
         are many plugins that aren't specifying a dependency and they blow up and runtime without p-u



git-svn-id: https://svn.apache.org/repos/asf/maven/components/branches/maven-2.0.x@522165 13f79535-47bb-0310-9956-ffa450edef68