PR: MNG-2339 - ${project.*} evaluated badly
This commit makes sure that pom.* and project.* are resolved using the model reflector,
but if the prefix isn't pom or project, context and model properties are consulted.
If the value still isn't found, the model reflector is used as a legacy fallback,
and a warning is printed.

Also, project.* is deprecated in favour of pom.*.

TODO: env.*.

N.B.: there's loads of warnings about ${artifactId}. I can turn these off,
but if you specify -DartifactId on the commandline, all hell breaks loose.

See http://docs.codehaus.org/display/MAVEN/Refactoring+Interpolation


git-svn-id: https://svn.apache.org/repos/asf/maven/components/trunk@549786 13f79535-47bb-0310-9956-ffa450edef68